requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
beartype==0.16.4
boolean.py==4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.5
cyclonedx-python-lib==5.2.0
defusedxml==0.7.1
Deprecated==1.2.14
docker==6.1.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.40
idna==3.4
isodate==0.6.1
license-expression==30.2.0
multidict==6.0.4
packageurl-python==0.11.2
packaging==23.2
ply==3.11
py-serializable==0.15.0
pycares==4.4.0
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
spdx-tools==0.8.2
typing_extensions==4.8.0
uritools==4.0.2
urllib3==2.0.7
websocket-client==1.6.4
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.3
